+++ arbitragelab/stochastic_control_approach/__init__.py
@@ -4,3 +4,4 @@ This module implements Stochastic Control Approach based Statistical Arbitrage s
 
 from arbitragelab.stochastic_control_approach.ou_model_jurek import OUModelJurek
 from arbitragelab.stochastic_control_approach.ou_model_mudchanatongsuk import OUModelMudchanatongsuk
+from arbitragelab.stochastic_control_approach.optimal_convergence import OptimalConvergence